Changed: double/Double to BigDecimal

+++ src/main/java/org/bilanzius/persistence/sql/adapter/SqlBankAccountAdapter.java
@@ -12,7 +12,7 @@ public class SqlBankAccountAdapter implements SqlDataAdapter<BankAccount>{
                 resultSet.getInt("id"),
                 resultSet.getInt("userId"),
                 resultSet.getString("name"),
-                resultSet.getDouble("balance")
+                resultSet.getBigDecimal("balance")
         );
     }
 }
